refactor(packages/eg-walker): split large test files into focused modules
Redistribute 5 oversized test files (ranging from 447–1,454 lines) into
12 additional focused files by subject under test. Test count goes from
30 to 42 files; all 305 tests pass. One typecheck error (widened
operation type in event-graph-branch-order) fixed post-split.

Before → after line counts:
  eg-walker-engine.test.ts   1454 → 444
  event-graph.test.ts        1368 → 549
  convergence-property.test.ts 1046 → 424
  replica.test.ts             867 → 239
  traversal-order-independence.test.ts 447 → 383

New files: replay-walker, indexed-sequence, partial-replay,
columnar-codec, event-graph-branch-order, event-graph-serialization,
event-graph-diff-versions, concurrent-deletes, surrogate-pair-convergence,
realistic-editing-traces, replica-advanced, compare-event-ids.
