+++ packages/eg-walker/src/test/convergence-property.test.ts
@@ -422,625 +422,3 @@ describe("EgWalkerReplica randomized convergence", () => {
     expect(failures, failures.slice(0, 3).join("\n")).toEqual([]);
   }, 60_000);
 });
-
-describe("EgWalkerReplica overlapping concurrent deletes", () => {
-  /**
-   * Two replicas delete partially overlapping ranges of the same
-   * shared text. The strong list-spec says the union of deletions
-   * survives in either traversal order; the engine's tombstone
-   * tracking must therefore make `everDeleted` idempotent across
-   * concurrent overlapping deletes.
-   */
-  const makeOverlappingDeleteEvents = (params: {
-    readonly base: string;
-    readonly aRange: readonly [number, number];
-    readonly bRange: readonly [number, number];
-  }): GraphEvent[] => {
-    const { base, aRange, bRange } = params;
-    return [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: base },
-        timestamp: 0,
-      },
-      {
-        id: "alice:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: {
-          type: OPERATION_TYPE.DELETE,
-          index: aRange[0],
-          length: aRange[1],
-        },
-        timestamp: 1,
-      },
-      {
-        id: "bob:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: {
-          type: OPERATION_TYPE.DELETE,
-          index: bRange[0],
-          length: bRange[1],
-        },
-        timestamp: 2,
-      },
-    ];
-  };
-
-  it("collapses fully-overlapping concurrent deletes to a single deletion", () => {
-    // Alice and Bob both delete the exact same 3-character range.
-    // The expected text is the survivor characters either side.
-    const events = makeOverlappingDeleteEvents({
-      base: "abcdefghij",
-      aRange: [2, 3],
-      bRange: [2, 3],
-    });
-    const text = canonicalText(events);
-    expect(text).toBe("abfghij");
-
-    // Either delivery order yields the same text.
-    const rand = createPrng(0x1010_1010);
-    for (let trial = 0; trial < 10; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText()).toBe(text);
-    }
-  });
-
-  it("merges partially-overlapping concurrent deletes (union of ranges)", () => {
-    // Alice deletes "cd" at index 2; Bob deletes "de" at index 3.
-    // Their union is "cde" — the resulting text is "abfghij".
-    const events = makeOverlappingDeleteEvents({
-      base: "abcdefghij",
-      aRange: [2, 2],
-      bRange: [3, 2],
-    });
-    const text = canonicalText(events);
-    expect(text).toBe("abfghij");
-
-    const rand = createPrng(0x2020_2020);
-    for (let trial = 0; trial < 10; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText()).toBe(text);
-    }
-  });
-
-  it("converges on randomized overlapping delete swarms", () => {
-    // Eight replicas, each repeatedly deleting random slices of a
-    // shared base string. The delete ranges drift over time as the
-    // local text shrinks, so concurrent deletes overlap in messy
-    // ways. The convergence guarantee says every replica still
-    // ends with the same text.
-    const REPLICA_COUNT = 6;
-    const baseText = "the quick brown fox jumps over the lazy dog";
-    const rootEvent: GraphEvent = {
-      id: "root:0",
-      parentVersion: new Set<EventId>(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: baseText },
-      timestamp: 0,
-    };
-
-    const allEvents: GraphEvent[] = [rootEvent];
-    const rand = createPrng(0xc0de_d00d);
-    // Each replica issues a few independent deletes against the
-    // *initial* base string (everyone forks off root:0), so all the
-    // deletes are pairwise concurrent.
-    for (let r = 0; r < REPLICA_COUNT; r++) {
-      const replicaId = `r${r}`;
-      for (let i = 0; i < 4; i++) {
-        const start = Math.floor(rand() * baseText.length);
-        const len = 1 + Math.floor(rand() * 6);
-        const safeLen = Math.min(baseText.length - start, len);
-        if (safeLen <= 0) {
-          continue;
-        }
-        // Manually craft the event so every delete forks off root:0
-        // (rather than off the replica's own previous delete) —
-        // that's what makes them pairwise concurrent.
-        const event: GraphEvent = {
-          id: `${replicaId}:${i}`,
-          parentVersion: new Set<EventId>(["root:0"]),
-          operation: {
-            type: OPERATION_TYPE.DELETE,
-            index: start,
-            length: safeLen,
-          },
-          timestamp: 1 + i,
-        };
-        allEvents.push(event);
-      }
-    }
-
-    const text = canonicalText(allEvents);
-    for (let trial = 0; trial < 12; trial++) {
-      const replica = applyInRandomDeliveryOrder(
-        `verify-${trial}`,
-        allEvents,
-        rand,
-      );
-      expect(replica.getText(), `trial=${trial}`).toBe(text);
-    }
-  });
-});
-
-describe("EgWalkerReplica concurrent multi-character inserts", () => {
-  it("never interleaves concurrent multi-character inserts", () => {
-    // Four authors all insert a 4-character word concurrently at the
-    // start of the same root text. The result must be the four words
-    // in some YATA-tie-broken order, none of them split across each
-    // other.
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "_" },
-        timestamp: 0,
-      },
-    ];
-    const words = ["alpha", "bravo", "delta", "gamma"];
-    for (let i = 0; i < words.length; i++) {
-      events.push({
-        id: `r${i}:0`,
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: words[i]! },
-        timestamp: 1 + i,
-      });
-    }
-    const text = canonicalText(events);
-
-    // Every word must survive contiguously.
-    for (const word of words) {
-      expect(text.includes(word), `word=${word}`).toBe(true);
-    }
-    // Anchor character is at the end (every concurrent insert lands
-    // at index 0 of the root text).
-    expect(text.endsWith("_")).toBe(true);
-
-    // Convergence across delivery orders.
-    const rand = createPrng(0x4040_4040);
-    for (let trial = 0; trial < 10; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText()).toBe(text);
-    }
-  });
-
-  it("preserves multi-character inserts as contiguous runs under random delivery", () => {
-    // Three replicas concurrently insert a 5-character word at
-    // random positions in a shared base text. The merged result
-    // must preserve each word as a contiguous substring.
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: {
-          type: OPERATION_TYPE.INSERT,
-          index: 0,
-          text: "0123456789",
-        },
-        timestamp: 0,
-      },
-    ];
-    const inserts: ReadonlyArray<{ id: string; at: number; text: string }> = [
-      { id: "alice:0", at: 0, text: "ALICE" },
-      { id: "bob:0", at: 5, text: "BOBOB" },
-      { id: "carol:0", at: 10, text: "CAROL" },
-    ];
-    for (let i = 0; i < inserts.length; i++) {
-      const { id, at, text } = inserts[i]!;
-      events.push({
-        id,
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: at, text },
-        timestamp: 1 + i,
-      });
-    }
-    const text = canonicalText(events);
-    for (const { text: word } of inserts) {
-      expect(text.includes(word), `word=${word}`).toBe(true);
-    }
-
-    const rand = createPrng(0x5050_5050);
-    for (let trial = 0; trial < 8; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText()).toBe(text);
-    }
-  });
-});
-
-describe("EgWalkerReplica surrogate-pair boundary cases", () => {
-  // A handful of non-BMP code points that the engine should always
-  // treat as 2-code-unit atomic units. The convergence property must
-  // hold even when an insert sits next to a surrogate pair, and the
-  // resulting text must remain well-formed UTF-16 (no lone
-  // surrogates).
-  const SURROGATE_EMOJI = "\u{1F600}"; // grinning face, UTF-16 length 2
-  const HEART = "\u{1F49C}"; // purple heart
-  const FLAG = "\u{1F1F8}\u{1F1EA}"; // regional indicator pair (Swedish flag)
-
-  const wellFormedUtf16 = (text: string): boolean => {
-    for (let i = 0; i < text.length; i++) {
-      const code = text.charCodeAt(i);
-      if (code >= 0xd800 && code <= 0xdbff) {
-        const next = i + 1 < text.length ? text.charCodeAt(i + 1) : 0;
-        if (next < 0xdc00 || next > 0xdfff) {
-          return false;
-        }
-        i++;
-      } else if (code >= 0xdc00 && code <= 0xdfff) {
-        return false;
-      }
-    }
-    return true;
-  };
-
-  it("rejects inserts that land between the halves of a surrogate pair", () => {
-    const replica = new EgWalkerReplica("r1", SURROGATE_EMOJI);
-    // Index 1 sits between the high and low surrogate of the emoji.
-    expect(() => replica.insert(1, "x")).toThrow(/surrogate halves/);
-    // Inserts at the BMP-aligned boundaries (0 and 2) are fine.
-    expect(() => replica.insert(0, "x")).not.toThrow();
-    expect(() => replica.insert(replica.getText().length, "y")).not.toThrow();
-    expect(wellFormedUtf16(replica.getText())).toBe(true);
-  });
-
-  it("converges across concurrent inserts around a surrogate pair", () => {
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: {
-          type: OPERATION_TYPE.INSERT,
-          index: 0,
-          text: `pre${SURROGATE_EMOJI}post`,
-        },
-        timestamp: 0,
-      },
-      {
-        id: "alice:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        // Insert immediately before the emoji (code-unit index 3).
-        operation: { type: OPERATION_TYPE.INSERT, index: 3, text: "A" },
-        timestamp: 1,
-      },
-      {
-        id: "bob:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        // Insert immediately after the emoji (code-unit index 5).
-        operation: { type: OPERATION_TYPE.INSERT, index: 5, text: "B" },
-        timestamp: 2,
-      },
-      {
-        id: "carol:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        // Insert another surrogate pair right after the existing one.
-        operation: { type: OPERATION_TYPE.INSERT, index: 5, text: HEART },
-        timestamp: 3,
-      },
-    ];
-
-    const text = canonicalText(events);
-    // Both concurrent atomic inserts and the surrogate-pair-inside
-    // edit must produce well-formed UTF-16.
-    expect(wellFormedUtf16(text)).toBe(true);
-    expect(text.includes(SURROGATE_EMOJI)).toBe(true);
-    expect(text.includes(HEART)).toBe(true);
-    expect(text.includes("A")).toBe(true);
-    expect(text.includes("B")).toBe(true);
-
-    const rand = createPrng(0x6060_6060);
-    for (let trial = 0; trial < 8; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText(), `trial=${trial}`).toBe(text);
-      expect(wellFormedUtf16(replica.getText())).toBe(true);
-    }
-  });
-
-  it("converges across concurrent deletes that wrap a surrogate pair", () => {
-    // Alice deletes a range that includes the whole emoji; Bob
-    // deletes a smaller subset on the same side. Both must agree on
-    // the result and never produce a lone surrogate.
-    const baseText = `xy${FLAG}zw`;
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: baseText },
-        timestamp: 0,
-      },
-      {
-        id: "alice:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        // Delete "y" + the regional-indicator pair (code-unit length 5).
-        operation: { type: OPERATION_TYPE.DELETE, index: 1, length: 5 },
-        timestamp: 1,
-      },
-      {
-        id: "bob:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        // Delete just "y".
-        operation: { type: OPERATION_TYPE.DELETE, index: 1, length: 1 },
-        timestamp: 2,
-      },
-    ];
-
-    const text = canonicalText(events);
-    expect(wellFormedUtf16(text)).toBe(true);
-    // Alice's larger delete dominates: only the trailing "zw"
-    // survives because the union of the two deletions removes the
-    // whole prefix `y${FLAG}`.
-    expect(text).toBe("xzw");
-
-    const rand = createPrng(0x7070_7070);
-    for (let trial = 0; trial < 6; trial++) {
-      const replica = applyInRandomDeliveryOrder(`r-${trial}`, events, rand);
-      expect(replica.getText(), `trial=${trial}`).toBe(text);
-      expect(wellFormedUtf16(replica.getText())).toBe(true);
-    }
-  });
-});
-
-describe("EgWalkerReplica realistic editing traces", () => {
-  /**
-   * Single-author trace that imitates "type a sentence, backspace,
-   * retype the end". No concurrent branches, but exercises the
-   * incremental-apply path repeatedly and round-trips through the
-   * (un)serializer.
-   */
-  it("round-trips a realistic single-author editing session", () => {
-    const replica = new EgWalkerReplica("solo");
-    replica.insert(0, "the quik brown fox");
-    // Realise we misspelled "quick".
-    replica.delete(4, 4); // remove "quik"
-    replica.insert(4, "quick"); // type it correctly
-    // Add a clause to the end.
-    replica.insert(replica.getText().length, " jumps");
-    // Backspace + retype the verb.
-    replica.delete(replica.getText().length - 6, 6);
-    replica.insert(replica.getText().length, " leaps");
-    expect(replica.getText()).toBe("the quick brown fox leaps");
-
-    // Serialize + deserialize and confirm we can keep editing.
-    const serialized = replica.serialize();
-    const restored = EgWalkerReplica.deserialize(serialized, "solo");
-    expect(restored.getText()).toBe("the quick brown fox leaps");
-    restored.insert(restored.getText().length, "!");
-    expect(restored.getText()).toBe("the quick brown fox leaps!");
-
-    // Replay counters should reflect the incremental path: no full
-    // replays after the cold start, every local edit advances the
-    // engine in place.
-    const stats = replica.getReplayStats();
-    expect(stats.incrementalApplies).toBeGreaterThan(0);
-    // The cold-start replay is the only full replay in a single-author
-    // trace.
-    expect(stats.fullReplays).toBeLessThanOrEqual(1);
-  });
-
-  /**
-   * Two authors collaborate on a shared draft. Each takes turns
-   * adding a sentence, with periodic syncs. The merged text must
-   * preserve every sentence the authors contributed.
-   */
-  it("merges a two-author collaborative draft deterministically", () => {
-    // Both replicas start blank; the shared `seed:0` event sets up
-    // the initial "Start. " prefix so every subsequent event has a
-    // common causal ancestor.
-    const alice = new EgWalkerReplica("alice");
-    const bob = new EgWalkerReplica("bob");
-    const seed: GraphEvent = {
-      id: "seed:0",
-      parentVersion: new Set<EventId>(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "Start. " },
-      timestamp: 0,
-    };
-    alice.applyRemoteEvent(cloneEvent(seed));
-    bob.applyRemoteEvent(cloneEvent(seed));
-
-    // Sync after every "turn" so we get a clean line of incremental
-    // applies rather than divergent forks. Two sync rounds.
-    const sync = (): void => {
-      const events: GraphEvent[] = [
-        ...alice.exportEventGraph().map(cloneEvent),
-        ...bob.exportEventGraph().map(cloneEvent),
-      ];
-      // `createPrng` is intentionally re-seeded with the *same*
-      // constant on every call so each sync round shuffles the
-      // event list in the same order. We want this test to assert
-      // that two authors converge under a specific (reproducible)
-      // delivery order; the dedicated "randomized convergence"
-      // tests above exercise the cross-seed delivery-order space.
-      const rand = createPrng(0x9999_9999);
-      for (const event of shuffled(events, rand)) {
-        alice.applyRemoteEvent(cloneEvent(event));
-        bob.applyRemoteEvent(cloneEvent(event));
-      }
-    };
-
-    alice.insert(alice.getText().length, "Alice writes intro. ");
-    bob.insert(bob.getText().length, "Bob outlines plan. ");
-    sync();
-    expect(alice.getText()).toBe(bob.getText());
-
-    alice.insert(alice.getText().length, "Alice expands intro. ");
-    bob.insert(bob.getText().length, "Bob adds detail. ");
-    sync();
-    expect(alice.getText()).toBe(bob.getText());
-
-    const merged = alice.getText();
-    expect(merged.includes("Alice writes intro.")).toBe(true);
-    expect(merged.includes("Bob outlines plan.")).toBe(true);
-    expect(merged.includes("Alice expands intro.")).toBe(true);
-    expect(merged.includes("Bob adds detail.")).toBe(true);
-
-    // Replay convergence: feeding both replicas' event graphs into a
-    // fresh replica under any delivery order yields the same text.
-    const allEvents: GraphEvent[] = [...alice.exportEventGraph()].map(
-      cloneEvent,
-    );
-    const rand = createPrng(0xaaaa_bbbb);
-    for (let trial = 0; trial < 8; trial++) {
-      const replica = applyInRandomDeliveryOrder(
-        `verify-${trial}`,
-        allEvents,
-        rand,
-      );
-      expect(replica.getText(), `trial=${trial}`).toBe(merged);
-    }
-  });
-
-  it("converges replicas that observe disjoint subsets of events first", () => {
-    // Three replicas, each seeing a *different* random ~60% subset
-    // of the event set before the rest is delivered. This is the
-    // canonical partition-then-heal pattern (network partition →
-    // partial catch-up → full reconciliation): every replica
-    // observes a distinct prefix, so during the first phase the
-    // replicas are genuinely out of sync, and convergence is only
-    // re-established after the second phase delivers the
-    // complement.
-    const { events, finalText } = runRandomizedMultiReplicaTrace({
-      replicaCount: 3,
-      eventBudget: 90,
-      seed: 0xcafe_babe,
-      maxInsertLen: 4,
-      deleteProbability: 0.3,
-      syncEveryN: 6,
-    });
-
-    const rand = createPrng(0xfafa_fafa);
-    const replicaCount = 3;
-    const replicas: EgWalkerReplica[] = [];
-    // Each replica gets its OWN random partition so the "first
-    // halves" the replicas observe really are disjoint subsets of
-    // the trace, not the same subset under different delivery
-    // orders.
-    const firstHalves: GraphEvent[][] = [];
-    const secondHalves: GraphEvent[][] = [];
-    for (let i = 0; i < replicaCount; i++) {
-      replicas.push(new EgWalkerReplica(`partition-${i}`));
-      const perReplica = shuffled(events.map(cloneEvent), rand);
-      const splitAt = Math.floor(perReplica.length * 0.6);
-      firstHalves.push(perReplica.slice(0, splitAt));
-      secondHalves.push(perReplica.slice(splitAt));
-    }
-
-    // Phase 1: each replica receives its own ~60% subset. After
-    // this phase the replicas have observed different subsets and
-    // may have buffered events whose causal predecessors are still
-    // in their respective second halves.
-    for (let i = 0; i < replicaCount; i++) {
-      for (const event of firstHalves[i]!) {
-        replicas[i]!.applyRemoteEvent(event);
-      }
-    }
-    // Sanity: at least one replica must see a partition that
-    // differs from at least one other replica's. If they were all
-    // identical we would not actually be testing the partition
-    // pattern — we would just be testing delivery-order variation,
-    // which is already covered by other tests in this file.
-    const firstHalfIdSets = firstHalves.map(
-      (half) => new Set<EventId>(half.map((event) => event.id)),
-    );
-    let sawDisjointPair = false;
-    for (let a = 0; a < replicaCount && !sawDisjointPair; a++) {
-      for (let b = a + 1; b < replicaCount; b++) {
-        const aSet = firstHalfIdSets[a]!;
-        const bSet = firstHalfIdSets[b]!;
-        for (const id of aSet) {
-          if (!bSet.has(id)) {
-            sawDisjointPair = true;
-            break;
-          }
-        }
-        if (sawDisjointPair) {
-          break;
-        }
-      }
-    }
-    expect(sawDisjointPair).toBe(true);
-
-    // Phase 2: deliver the complement. After this every replica
-    // has observed every event (in some delivery order) and must
-    // converge to the canonical text.
-    for (let i = 0; i < replicaCount; i++) {
-      for (const event of secondHalves[i]!) {
-        replicas[i]!.applyRemoteEvent(event);
-      }
-      expect(replicas[i]!.getPendingRemoteCount()).toBe(0);
-      expect(replicas[i]!.getText()).toBe(finalText);
-    }
-  });
-});
-
-describe("EgWalkerReplica causal-buffering under random delivery", () => {
-  it("buffers and flushes out-of-order remote events without diverging", () => {
-    // Build a forked DAG: root → A → C; root → B → D; D parents
-    // include both A and B (deep merge). Then deliver the events to
-    // a fresh replica in reverse order so every event lands before
-    // its parents and gets buffered, exercising the
-    // `pendingByMissingParent` path on every event except root.
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set<EventId>(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "ROOT" },
-        timestamp: 0,
-      },
-      {
-        id: "A:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 4, text: "-A" },
-        timestamp: 1,
-      },
-      {
-        id: "B:0",
-        parentVersion: new Set<EventId>(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "B-" },
-        timestamp: 2,
-      },
-      {
-        id: "C:0",
-        parentVersion: new Set<EventId>(["A:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "C-" },
-        timestamp: 3,
-      },
-      {
-        id: "D:0",
-        parentVersion: new Set<EventId>(["A:0", "B:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "D-" },
-        timestamp: 4,
-      },
-    ];
-    const expected = canonicalText(events);
-
-    // Reverse delivery: every event arrives before its parents.
-    const replica = new EgWalkerReplica("reverse");
-    for (let i = events.length - 1; i >= 0; i--) {
-      replica.applyRemoteEvent(cloneEvent(events[i]!));
-    }
-    expect(replica.getPendingRemoteCount()).toBe(0);
-    expect(replica.getText()).toBe(expected);
-  });
-
-  it("converges with replicas receiving completely shuffled deliveries", () => {
-    // A larger sweep across random deliveries. Each trial picks a
-    // random permutation, drops the events through a fresh replica,
-    // and asserts the final text matches the canonical replay.
-    const { events, finalText } = runRandomizedMultiReplicaTrace({
-      replicaCount: 4,
-      eventBudget: 60,
-      seed: 0x5555_6666,
-      maxInsertLen: 3,
-      deleteProbability: 0.25,
-      syncEveryN: 5,
-    });
-
-    const rand = createPrng(0x7777_8888);
-    for (let trial = 0; trial < 20; trial++) {
-      const replica = applyInRandomDeliveryOrder(
-        `trial-${trial}`,
-        events,
-        rand,
-      );
-      expect(replica.getText(), `trial=${trial}`).toBe(finalText);
-    }
-  });
-});
