+++ packages/eg-walker/src/test/eg-walker-engine.test.ts
@@ -1,31 +1,9 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { OPERATION_TYPE } from "../constants/operation-types";
-import lz4 from "lz4js";
-import { CriticalVersionAnalyzer } from "../engine/critical-version";
-import { ReplayWalker } from "../core/replay-walker";
 import { EgWalkerReplica } from "../core/replica";
 import { EgWalkerEngine } from "../engine/eg-walker-engine";
 import { EventGraph } from "../graph/event-graph";
-import { PartialReplayManager } from "../engine/partial-replay";
-import { IndexedSequence } from "../engine/indexed-sequence";
-import { ColumnarEventGraphCodec } from "../graph/columnar-codec";
 import type { EventId, GraphEvent } from "../types";
-import { createPrng } from "./test-helpers";
-
-interface SequenceModelItem {
-  readonly id: string;
-  prepare: number;
-  effect: number;
-}
-
-const visiblePositions = (
-  items: ReadonlyArray<SequenceModelItem>,
-  kind: "prepare" | "effect",
-): number[] =>
-  items
-    .map((item, index) => ({ item, index }))
-    .filter(({ item }) => item[kind] === 1)
-    .map(({ index }) => index);
 
 describe("EgWalkerEngine", () => {
   it("transforms concurrent insertions from the paper's motivating example", () => {
@@ -227,85 +205,6 @@ describe("EgWalkerEngine", () => {
   });
 });
 
-describe("ReplayWalker", () => {
-  it("walks events through the engine and exposes final versions", () => {
-    const walker = new ReplayWalker({ initialText: "Hi" });
-    const result = walker.walk([
-      {
-        id: "alice:0",
-        parentVersion: new Set(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 2, text: "!" },
-        timestamp: 1,
-      },
-      {
-        id: "alice:1",
-        parentVersion: new Set(["alice:0"]),
-        operation: { type: OPERATION_TYPE.DELETE, index: 0, length: 1 },
-        timestamp: 2,
-      },
-    ]);
-
-    expect(result.finalText).toBe("i!");
-    expect(result.eventsProcessed).toBe(2);
-    expect(walker.getPrepareVersion()).toEqual(new Set(["alice:1"]));
-    expect(walker.getEffectVersion()).toEqual(new Set(["alice:1"]));
-  });
-
-  it("walks unordered complete event batches", () => {
-    const walker = new ReplayWalker();
-    const result = walker.walk([
-      {
-        id: "alice:1",
-        parentVersion: new Set(["alice:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 1, text: "B" },
-        timestamp: 2,
-      },
-      {
-        id: "alice:0",
-        parentVersion: new Set(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-        timestamp: 1,
-      },
-    ]);
-
-    expect(result.finalText).toBe("AB");
-    expect(result.eventsProcessed).toBe(2);
-    expect(walker.getPrepareVersion()).toEqual(new Set(["alice:1"]));
-  });
-
-  it("walks an empty event list without changing initial text", () => {
-    const walker = new ReplayWalker({ initialText: "seed" });
-
-    expect(walker.walk([])).toEqual({
-      finalText: "seed",
-      eventsProcessed: 0,
-      retreatCount: 0,
-      advanceCount: 0,
-      nonConflictingRunCount: 0,
-      fullReplayCount: 0,
-      // The engine seeds the initial-text placeholder during `reset` and
-      // samples the peak right after, so even an empty event list reports
-      // 1 live record.
-      peakSequenceRecordCount: 1,
-    });
-    expect(walker.getPrepareVersion()).toEqual(new Set());
-    expect(walker.getEffectVersion()).toEqual(new Set());
-  });
-
-  it("defaults walker initial text to an empty string", () => {
-    const result = new ReplayWalker().walk([
-      {
-        id: "alice:0",
-        parentVersion: new Set(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-        timestamp: 1,
-      },
-    ]);
-
-    expect(result.finalText).toBe("A");
-  });
-});
-
 describe("EventGraph version semantics", () => {
   it("expands frontier versions and diffs transitive event sets", () => {
     const graph = new EventGraph();
@@ -340,915 +239,6 @@ describe("EventGraph version semantics", () => {
       onlyInLeft: new Set(["left"]),
       onlyInRight: new Set(["right"]),
     });
-  });
-});
-
-describe("Full paper architecture utilities", () => {
-  it("maps prepare/effect indexes through the ranked B-tree sequence", () => {
-    const items = [
-      { id: "a", prepare: 1, effect: 1 },
-      { id: "b", prepare: 0, effect: 1 },
-      { id: "c", prepare: 1, effect: 0 },
-      { id: "d", prepare: 1, effect: 1 },
-    ];
-    const sequence = new IndexedSequence(
-      (item: (typeof items)[number]) => item.prepare,
-      (item: (typeof items)[number]) => item.effect,
-      items,
-    );
-
-    expect(sequence.prepareIndexToPosition(0, false)).toBe(0);
-    expect(sequence.prepareIndexToPosition(1, false)).toBe(2);
-    expect(sequence.nextPrepareVisiblePosition(1)).toBe(2);
-    expect(sequence.effectIndexBeforePosition(3)).toBe(2);
-
-    items[1]!.prepare = 1;
-    items[2]!.effect = 1;
-    sequence.updateItem(items[1]!);
-    sequence.updateItem(items[2]!);
-
-    expect(sequence.prepareIndexToPosition(1, false)).toBe(1);
-    expect(sequence.effectIndexBeforePosition(4)).toBe(4);
-  });
-
-  it("distinguishes out-of-range lookups from structural errors via tryPrepareIndexToPositionAndOffset", () => {
-    const items = [
-      { id: "a", prepare: 1, effect: 1 },
-      { id: "b", prepare: 0, effect: 1 },
-      { id: "c", prepare: 1, effect: 0 },
-    ];
-    const sequence = new IndexedSequence(
-      (item: (typeof items)[number]) => item.prepare,
-      (item: (typeof items)[number]) => item.effect,
-      items,
-    );
-
-    // Within range: prepare-index 1 lands at the second prepare-visible
-    // record (position 2, which is item "c") with no record offset.
-    expect(sequence.tryPrepareIndexToPositionAndOffset(1, false)).toEqual({
-      position: 2,
-      offsetInRecord: 0,
-    });
-
-    // Past the prepare-visible weight sum and negative indexes return
-    // `undefined` instead of throwing, so callers can distinguish the
-    // expected end-of-text condition from real bugs without a
-    // catch-all try/catch.
-    expect(
-      sequence.tryPrepareIndexToPositionAndOffset(2, false),
-    ).toBeUndefined();
-    expect(
-      sequence.tryPrepareIndexToPositionAndOffset(-1, false),
-    ).toBeUndefined();
-
-    // The throwing variant still raises on the same inputs so structural
-    // bugs (aggregate corruption etc.) propagate as before.
-    expect(() => sequence.prepareIndexToPositionAndOffset(2, false)).toThrow(
-      /out of bounds/,
-    );
-  });
-
-  it("keeps ranked B-tree indexes correct across leaf and internal splits", () => {
-    const items = Array.from({ length: 2_200 }, (_, index) => ({
-      id: `item-${index}`,
-      prepare: index % 3 === 0 ? 0 : 1,
-      effect: index % 5 === 0 ? 0 : 1,
-    }));
-    const sequence = new IndexedSequence(
-      (item: (typeof items)[number]) => item.prepare,
-      (item: (typeof items)[number]) => item.effect,
-    );
-
-    for (const item of items) {
-      sequence.push(item);
-    }
-
-    const expectedPreparePositions = items
-      .map((item, index) => ({ item, index }))
-      .filter(({ item }) => item.prepare === 1)
-      .map(({ index }) => index);
-    const expectedEffectBefore1_700 = items
-      .slice(0, 1_700)
-      .filter((item) => item.effect === 1).length;
-
-    expect(sequence.length).toBe(items.length);
-    expect(sequence.at(1_799)).toBe(items[1_799]);
-    expect(sequence.positionOf(items[1_799]!)).toBe(1_799);
-    expect(sequence.prepareIndexToPosition(40, false)).toBe(
-      expectedPreparePositions[40],
-    );
-    expect(sequence.effectIndexBeforePosition(1_700)).toBe(
-      expectedEffectBefore1_700,
-    );
-
-    items[1_500]!.prepare = 1;
-    items[1_500]!.effect = 1;
-    sequence.updateItem(items[1_500]!);
-
-    expect(sequence.nextPrepareVisiblePosition(1_499)).toBe(1_499);
-    expect(sequence.positionOf(items[2_000]!)).toBe(2_000);
-  });
-
-  it("matches an array model across deterministic B-tree inserts and updates", () => {
-    const random = createPrng(13_337);
-    const model: SequenceModelItem[] = [];
-    const sequence = new IndexedSequence<SequenceModelItem>(
-      (item) => item.prepare,
-      (item) => item.effect,
-    );
-
-    for (let step = 0; step < 1_000; step++) {
-      const item: SequenceModelItem = {
-        id: `item-${step}`,
-        prepare: random() < 0.7 ? 1 : 0,
-        effect: random() < 0.8 ? 1 : 0,
-      };
-      const index = Math.floor(random() * (model.length + 1));
-      model.splice(index, 0, item);
-      sequence.insert(index, item);
-
-      if (step % 5 === 0 && model.length > 0) {
-        const updateIndex = Math.floor(random() * model.length);
-        const updated = model[updateIndex]!;
-        updated.prepare = updated.prepare === 1 ? 0 : 1;
-        updated.effect = updated.effect === 1 ? 0 : 1;
-        sequence.updateItem(updated);
-      }
-
-      const probeIndex = Math.floor(random() * model.length);
-      const probed = model[probeIndex]!;
-      expect(sequence.at(probeIndex)).toBe(probed);
-      expect(sequence.positionOf(probed)).toBe(probeIndex);
-
-      const preparePositions = visiblePositions(model, "prepare");
-      if (preparePositions.length > 0) {
-        const prepareIndex = Math.floor(random() * preparePositions.length);
-        expect(sequence.prepareIndexToPosition(prepareIndex, false)).toBe(
-          preparePositions[prepareIndex],
-        );
-      }
-
-      const effectProbe = Math.floor(random() * (model.length + 1));
-      expect(sequence.effectIndexBeforePosition(effectProbe)).toBe(
-        model.slice(0, effectProbe).filter((item) => item.effect === 1).length,
-      );
-    }
-
-    expect(sequence.toArray()).toEqual(model);
-  });
-
-  it("covers ranked B-tree boundary behavior and bulk weight refresh", () => {
-    const items = [
-      { id: "a", prepare: 1, effect: 0 },
-      { id: "b", prepare: 0, effect: 1 },
-    ];
-    const sequence = new IndexedSequence(
-      (item: (typeof items)[number]) => item.prepare,
-      (item: (typeof items)[number]) => item.effect,
-      items,
-    );
-
-    expect(sequence.at(-1)).toBeUndefined();
-    expect(sequence.at(2)).toBeUndefined();
-    expect(sequence.slice(1)).toEqual([items[1]]);
-    expect(sequence.indexOf((item) => item.id === "b")).toBe(1);
-    expect(sequence.indexOf((item) => item.id === "missing")).toBe(-1);
-    expect(sequence.positionOf({ id: "external", prepare: 1, effect: 1 })).toBe(
-      -1,
-    );
-    expect(sequence.nextPrepareVisiblePosition(-1)).toBeNull();
-    expect(sequence.nextPrepareVisiblePosition(3)).toBeNull();
-    expect(sequence.prepareIndexToPosition(1, true)).toBe(2);
-    expect(() => sequence.prepareIndexToPosition(-1, false)).toThrow(
-      "Index -1 out of bounds",
-    );
-    expect(() => sequence.prepareIndexToPosition(1, false)).toThrow(
-      "Index 1 out of bounds",
-    );
-    expect(() => sequence.insert(-1, items[0]!)).toThrow(
-      "Insert index -1 out of bounds",
-    );
-
-    items[0]!.prepare = 0;
-    items[0]!.effect = 1;
-    items[1]!.prepare = 1;
-    items[1]!.effect = 0;
-    sequence.updateWeights();
-
-    expect(sequence.prepareIndexToPosition(0, false)).toBe(1);
-    expect(sequence.effectIndexBeforePosition(2)).toBe(1);
-
-    sequence.clear();
-    expect(sequence.length).toBe(0);
-    expect(sequence.toArray()).toEqual([]);
-    expect(sequence.indexOf(() => true)).toBe(-1);
-    expect(sequence.effectIndexBeforePosition(10)).toBe(0);
-    expect(sequence.prepareIndexToPosition(0, true)).toBe(0);
-    expect(() => sequence.prepareIndexToPosition(0, false)).toThrow(
-      "Index 0 out of bounds",
-    );
-    expect(() => sequence.insert(1, items[0]!)).toThrow(
-      "Insert index 1 out of bounds",
-    );
-    expect(() => sequence.updateItem(items[0]!)).not.toThrow();
-  });
-
-  it("refreshes ranked B-tree weights across internal nodes", () => {
-    const items = Array.from({ length: 140 }, (_, index) => ({
-      id: `bulk-${index}`,
-      prepare: 1,
-      effect: 1,
-    }));
-    const sequence = new IndexedSequence(
-      (item: (typeof items)[number]) => item.prepare,
-      (item: (typeof items)[number]) => item.effect,
-      items,
-    );
-
-    items.forEach((item, index) => {
-      item.prepare = index === 139 ? 1 : 0;
-      item.effect = index === 0 ? 1 : 0;
-    });
-    sequence.updateWeights();
-
-    expect(sequence.prepareIndexToPosition(0, false)).toBe(139);
-    expect(sequence.effectIndexBeforePosition(140)).toBe(1);
-  });
-
-  it("partially replays from a critical checkpoint using the full graph", () => {
-    const graph = new EventGraph();
-    const events: GraphEvent[] = [
-      {
-        id: "r:0",
-        parentVersion: new Set(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "Hello" },
-        timestamp: 1,
-      },
-      {
-        id: "r:1",
-        parentVersion: new Set(["r:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 5, text: "!" },
-        timestamp: 2,
-      },
-    ];
-    events.forEach((event) => graph.addEvent(event));
-
-    const result = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: "Hello",
-    });
-
-    expect(result.text).toBe("Hello!");
-    expect(result.replayedEventIds).toEqual(["r:1"]);
-  });
-
-  it("splits a placeholder for an insert landing inside checkpoint-era text", () => {
-    // Section 3.6: pre-checkpoint content is one placeholder record. An insert
-    // at index 5 must split it into ["Hello"] + new chars + [" world"] without
-    // materializing one CRDT item per pre-checkpoint code unit.
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "r:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "Hello world" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "r:1",
-      parentVersion: new Set(["r:0"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 5, text: "!!" },
-      timestamp: 2,
-    });
-
-    const result = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: "Hello world",
-    });
-
-    expect(result.text).toBe("Hello!! world");
-    expect(result.replayedEventIds).toEqual(["r:1"]);
-  });
-
-  it("splits a placeholder for a delete touching checkpoint-era text", () => {
-    // Single-char delete inside the placeholder must split off exactly one
-    // character so prepare/effect indexing for the remaining region stays
-    // accurate.
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "r:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "abcdef" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "r:1",
-      parentVersion: new Set(["r:0"]),
-      operation: { type: OPERATION_TYPE.DELETE, index: 2, length: 1 },
-      timestamp: 2,
-    });
-
-    const result = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: "abcdef",
-    });
-
-    expect(result.text).toBe("abdef");
-    expect(result.replayedEventIds).toEqual(["r:1"]);
-  });
-
-  it("handles a multi-char delete inside a placeholder in one batch split", () => {
-    // The delete spans 3 code units of the placeholder. The engine should split
-    // the placeholder into [prefix, deleted-segment, suffix] in one shot rather
-    // than per-character, while still emitting a single transformed delete run.
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "r:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "abcdefghij" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "r:1",
-      parentVersion: new Set(["r:0"]),
-      operation: { type: OPERATION_TYPE.DELETE, index: 3, length: 4 },
-      timestamp: 2,
-    });
-
-    const result = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: "abcdefghij",
-    });
-
-    expect(result.text).toBe("abchij");
-    expect(result.transformedOperations).toEqual([
-      { type: OPERATION_TYPE.DELETE, index: 3, length: 4 },
-    ]);
-  });
-
-  it("supports interleaved inserts and deletes against placeholder text", () => {
-    // Mixed workload: insert then delete inside the placeholder region. The
-    // partial replay result must match what a full replay from scratch
-    // produces, proving that splitting placeholders preserves convergence.
-    const graph = new EventGraph();
-    const events: GraphEvent[] = [
-      {
-        id: "r:0",
-        parentVersion: new Set(),
-        operation: {
-          type: OPERATION_TYPE.INSERT,
-          index: 0,
-          text: "The quick brown fox",
-        },
-        timestamp: 1,
-      },
-      {
-        id: "r:1",
-        parentVersion: new Set(["r:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 10, text: "very " },
-        timestamp: 2,
-      },
-      {
-        id: "r:2",
-        parentVersion: new Set(["r:1"]),
-        operation: { type: OPERATION_TYPE.DELETE, index: 4, length: 6 },
-        timestamp: 3,
-      },
-    ];
-    events.forEach((event) => graph.addEvent(event));
-
-    const partial = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: "The quick brown fox",
-    });
-    const full = new EgWalkerEngine().generate(graph.getTopologicalOrder());
-
-    expect(partial.text).toBe(full.text);
-  });
-
-  it("converges with full replay for overlapping concurrent deletes inside placeholder text", () => {
-    // Regression: two concurrent multi-character deletes targeting the same
-    // checkpoint-era region used to mangle the document because the
-    // placeholder branch removed text again after a sibling delete had
-    // already retired the same segment. Partial replay must match full
-    // replay for both exact and partial overlaps.
-    const buildGraph = (
-      e2: GraphEvent["operation"],
-    ): { graph: EventGraph; events: GraphEvent[] } => {
-      const graph = new EventGraph();
-      const events: GraphEvent[] = [
-        {
-          id: "r:0",
-          parentVersion: new Set(),
-          operation: {
-            type: OPERATION_TYPE.INSERT,
-            index: 0,
-            text: "abcdef",
-          },
-          timestamp: 1,
-        },
-        {
-          id: "alice:0",
-          parentVersion: new Set(["r:0"]),
-          operation: { type: OPERATION_TYPE.DELETE, index: 2, length: 2 },
-          timestamp: 2,
-        },
-        {
-          id: "bob:0",
-          parentVersion: new Set(["r:0"]),
-          operation: e2,
-          timestamp: 3,
-        },
-      ];
-      events.forEach((event) => graph.addEvent(event));
-      return { graph, events };
-    };
-
-    const exact = buildGraph({
-      type: OPERATION_TYPE.DELETE,
-      index: 2,
-      length: 2,
-    });
-    const partial = buildGraph({
-      type: OPERATION_TYPE.DELETE,
-      index: 1,
-      length: 4,
-    });
-
-    for (const { graph } of [exact, partial]) {
-      const fullText = new EgWalkerEngine().generate(
-        graph.getTopologicalOrder(),
-      ).text;
-      const partialText = new PartialReplayManager().replayFromCheckpoint(
-        graph,
-        { version: new Set(["r:0"]), text: "abcdef" },
-      ).text;
-      expect(partialText).toBe(fullText);
-    }
-  });
-
-  it("converges with full replay when concurrent inserts split the same placeholder at different offsets", () => {
-    // Regression for partial-replay placeholder splitting: when
-    // partial replay starts from a checkpoint, pre-checkpoint text is
-    // collapsed into a single placeholder record. Two concurrent
-    // inserts inside that placeholder split it at different offsets,
-    // and a descendant of one of those inserts is later integrated
-    // into the conflict region created by the other. The YATA
-    // integration scan compares origin ids by identity, so the engine
-    // has to rewrite the existing items' `originLeft` references when
-    // a placeholder splits — otherwise siblings anchored to the same
-    // logical boundary look as if they have different origins and
-    // partial replay diverges from full replay.
-    const graph = new EventGraph();
-    const events: GraphEvent[] = [
-      {
-        id: "root:0",
-        parentVersion: new Set(),
-        operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "abcdef" },
-        timestamp: 1,
-      },
-      {
-        id: "b:0",
-        parentVersion: new Set(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 2, text: "B" },
-        timestamp: 2,
-      },
-      {
-        id: "c:0",
-        parentVersion: new Set(["root:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 1, text: "C" },
-        timestamp: 3,
-      },
-      {
-        id: "f:0",
-        parentVersion: new Set(["c:0"]),
-        operation: { type: OPERATION_TYPE.INSERT, index: 3, text: "F" },
-        timestamp: 4,
-      },
-    ];
-    events.forEach((event) => graph.addEvent(event));
-
-    const fullText = new EgWalkerEngine().generate(
-      graph.getTopologicalOrder(),
-    ).text;
-    const partialText = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["root:0"]),
-      text: "abcdef",
-    }).text;
-
-    expect(partialText).toBe(fullText);
-    expect(partialText).toBe("aCbBFcdef");
-  });
-
-  it("does not materialize per-character CRDT items for the checkpoint text", () => {
-    // Acceptance criterion for issue #665: partial replay must avoid the
-    // O(checkpoint.length) item cost. We use a long pre-checkpoint string and
-    // a single trailing insert. The engine reports `eventsProcessed` covering
-    // only the divergent suffix; the placeholder representation is what makes
-    // this realistic for large documents.
-    const checkpointText = "x".repeat(5000);
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "r:0",
-      parentVersion: new Set(),
-      operation: {
-        type: OPERATION_TYPE.INSERT,
-        index: 0,
-        text: checkpointText,
-      },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "r:1",
-      parentVersion: new Set(["r:0"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 2500, text: "!" },
-      timestamp: 2,
-    });
-
-    const result = new PartialReplayManager().replayFromCheckpoint(graph, {
-      version: new Set(["r:0"]),
-      text: checkpointText,
-    });
-
-    expect(result.text).toBe(`${"x".repeat(2500)}!${"x".repeat(2500)}`);
-    expect(result.stats.eventsProcessed).toBe(1);
-  });
-
-  it("detects linear frontier versions as critical checkpoints", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "r:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "r:1",
-      parentVersion: new Set(["r:0"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 1, text: "B" },
-      timestamp: 2,
-    });
-
-    const analyzer = new CriticalVersionAnalyzer();
-
-    expect(analyzer.isCritical(graph, new Set(["r:0"]))).toBe(true);
-    expect(analyzer.latestCriticalVersion(graph)).toEqual(new Set(["r:1"]));
-  });
-
-  it("handles empty and non-critical versions in critical checkpoint analysis", () => {
-    const graph = new EventGraph();
-    const analyzer = new CriticalVersionAnalyzer();
-
-    expect(analyzer.isCritical(graph, new Set())).toBe(true);
-    expect(analyzer.latestCriticalVersion(graph)).toEqual(new Set());
-
-    graph.addEvent({
-      id: "left",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "L" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "right",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "R" },
-      timestamp: 2,
-    });
-
-    expect(analyzer.isCritical(graph, new Set(["left"]))).toBe(false);
-  });
-
-  it("rejects multi-frontier checkpoints with partially descended events", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "a",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "b",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "B" },
-      timestamp: 2,
-    });
-    graph.addEvent({
-      id: "c",
-      parentVersion: new Set(["a"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 1, text: "C" },
-      timestamp: 3,
-    });
-
-    const analyzer = new CriticalVersionAnalyzer();
-
-    expect(analyzer.isCritical(graph, new Set(["a", "b"]))).toBe(false);
-  });
-
-  it("round-trips the event graph through the columnar codec", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "alice:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "Hi" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "alice:1",
-      parentVersion: new Set(["alice:0"]),
-      operation: { type: OPERATION_TYPE.DELETE, index: 1, length: 1 },
-      timestamp: 2,
-    });
-
-    const codec = new ColumnarEventGraphCodec();
-    const decoded = codec.decode(codec.encode(graph));
-
-    expect(decoded.getTopologicalOrder()).toEqual(graph.getTopologicalOrder());
-    expect(decoded.getFrontier()).toEqual(new Set(["alice:1"]));
-  });
-
-  it("round-trips the event graph through the binary columnar codec", () => {
-    const graph = new EventGraph();
-    for (let i = 0; i < 20; i++) {
-      graph.addEvent({
-        id: `alice:${i}`,
-        parentVersion: i === 0 ? new Set() : new Set([`alice:${i - 1}`]),
-        operation: { type: OPERATION_TYPE.INSERT, index: i, text: "x" },
-        timestamp: 1_778_000_000_000 + i,
-      });
-    }
-
-    const codec = new ColumnarEventGraphCodec();
-    const decoded = codec.decodeBinary(codec.encodeBinary(graph));
-
-    expect(decoded.getTopologicalOrder()).toEqual(graph.getTopologicalOrder());
-    expect(
-      new EgWalkerEngine().generate(decoded.getTopologicalOrder()).text,
-    ).toBe("x".repeat(20));
-  });
-
-  it("preserves a single sequence-zero generated ID through columnar codecs", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "alice:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-      timestamp: 1,
-    });
-
-    const codec = new ColumnarEventGraphCodec();
-    const decodedText = codec.decode(codec.encode(graph));
-    const decodedBinary = codec.decodeBinary(codec.encodeBinary(graph));
-
-    expect(decodedText.getTopologicalOrder().map((event) => event.id)).toEqual([
-      "alice:0",
-    ]);
-    expect(
-      decodedBinary.getTopologicalOrder().map((event) => event.id),
-    ).toEqual(["alice:0"]);
-  });
-
-  it("decodes alternating columnar operation runs in linear run order", () => {
-    const graph = new EventGraph();
-    for (let i = 0; i < 800; i++) {
-      graph.addEvent({
-        id: `event:${i}`,
-        parentVersion: i === 0 ? new Set() : new Set([`event:${i - 1}`]),
-        operation:
-          i % 2 === 0
-            ? { type: OPERATION_TYPE.INSERT, index: 0, text: "x" }
-            : { type: OPERATION_TYPE.DELETE, index: 0, length: 1 },
-        timestamp: i,
-      });
-    }
-
-    const codec = new ColumnarEventGraphCodec();
-    const encoded = codec.encode(graph);
-    const decoded = codec.decode(encoded);
-
-    expect(encoded.operationRuns).toHaveLength(800);
-    expect(decoded.getTopologicalOrder()).toEqual(graph.getTopologicalOrder());
-  });
-
-  it("rejects malformed binary columnar payloads", () => {
-    const codec = new ColumnarEventGraphCodec();
-
-    expect(() => codec.decodeBinary(new Uint8Array())).toThrow(
-      "Unexpected end of varint",
-    );
-    expect(() => codec.decodeBinary(new Uint8Array([1, 0]))).toThrow(
-      "Invalid eg-walker columnar graph header",
-    );
-    expect(() => codec.decodeBinary(new Uint8Array([4, 0x45, 0x47]))).toThrow(
-      "Unexpected end of binary eg-walker graph",
-    );
-  });
-
-  it("covers columnar ID and malformed operation run edge cases", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "custom-id",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "A" },
-      timestamp: 1,
-    });
-    graph.addEvent({
-      id: "replica:not-number",
-      parentVersion: new Set(["custom-id"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 1, text: "B" },
-      timestamp: 2,
-    });
-
-    const codec = new ColumnarEventGraphCodec();
-    const encoded = codec.encode(graph);
-
-    expect(codec.toSerializedGraph(encoded).events).toHaveLength(2);
-    expect(
-      codec.decodeBinary(codec.encodeBinary(graph)).getAllEvents(),
-    ).toHaveLength(2);
-    expect(() =>
-      codec.decode({
-        ...encoded,
-        operationRuns: [],
-      }),
-    ).toThrow("Missing operation run for event offset 0");
-    expect(() =>
-      codec.decode({
-        ...encoded,
-        operationRuns: [
-          {
-            type: OPERATION_TYPE.INSERT,
-            startIndex: 0,
-            startEventOffset: 1,
-            length: 1,
-            textLength: 1,
-          },
-        ],
-      }),
-    ).toThrow("Operation run 0 does not cover event offset 0");
-
-    expect(() =>
-      codec.decode({
-        ...encoded,
-        parentOverrides: [{ eventOffset: 1, parents: ["missing"] }],
-      }),
-    ).toThrow("Missing parent event: missing");
-
-    graph.addEvent({
-      id: "replica:2",
-      parentVersion: new Set(["replica:not-number"]),
-      operation: { type: OPERATION_TYPE.INSERT, index: 2, text: "C" },
-      timestamp: Number.NaN,
-    });
-    // EGW3 zigzag-delta encodes timestamps so negative values are valid; the
-    // varint guard still rejects non-safe integers like NaN/Infinity.
-    expect(() => codec.encodeBinary(graph)).toThrow(
-      "Cannot encode invalid zigzag varint value NaN",
-    );
-  });
-
-  it("round-trips a large inserted-content payload through the binary codec", () => {
-    // Build distinct, mostly-incompressible blocks so LZ4 cannot dedupe across
-    // events. Exercises the BinaryWriter buffer growth that previously
-    // overflowed JS arg limits via push-spread of a Uint8Array.
-    let prng = 0x9e_37_79_b9 >>> 0;
-    const nextChar = (): string => {
-      prng = (prng * 1_103_515_245 + 12_345) >>> 0;
-      return String.fromCharCode(0x21 + (prng % 94));
-    };
-    const blocks = Array.from({ length: 32 }, () =>
-      Array.from({ length: 4_096 }, nextChar).join(""),
-    );
-
-    const graph = new EventGraph();
-    let cursor = 0;
-    blocks.forEach((block, i) => {
-      graph.addEvent({
-        id: `bulk:${i}`,
-        parentVersion: i === 0 ? new Set() : new Set([`bulk:${i - 1}`]),
-        operation: { type: OPERATION_TYPE.INSERT, index: cursor, text: block },
-        timestamp: 1_778_000_000_000 + i,
-      });
-      cursor += block.length;
-    });
-
-    const codec = new ColumnarEventGraphCodec();
-    const encoded = codec.encodeBinary(graph);
-    const decoded = codec.decodeBinary(encoded);
-    const text = new EgWalkerEngine().generate(
-      decoded.getTopologicalOrder(),
-    ).text;
-    expect(text).toBe(blocks.join(""));
-  });
-
-  it("caps lz4 destination allocation against the declared textLengths sum", () => {
-    const graph = new EventGraph();
-    graph.addEvent({
-      id: "bounded:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "😀" },
-      timestamp: 1,
-    });
-
-    const codec = new ColumnarEventGraphCodec();
-    const encoded = codec.encodeBinary(graph);
-    const decompressSpy = vi.spyOn(lz4, "decompress");
-
-    try {
-      codec.decodeBinary(encoded);
-      // 2 UTF-16 code units * 4 + 64 = 72 bytes maxInsertedBytes.
-      expect(decompressSpy).toHaveBeenCalledWith(expect.any(Uint8Array), 72);
-    } finally {
-      decompressSpy.mockRestore();
-    }
-  });
-
-  it("rejects payloads whose decompressed content does not match declared textLengths", () => {
-    const codec = new ColumnarEventGraphCodec();
-    // Real, well-formed payload: one 5-char insert. textLengths sum = 5.
-    const realGraph = new EventGraph();
-    realGraph.addEvent({
-      id: "real:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "hello" },
-      timestamp: 1,
-    });
-    const realPayload = codec.encodeBinary(realGraph);
-
-    // Tampered payload: a 1-char insert (textLengths sum = 1) but the LZ4
-    // frame still carries the 5-byte original content. After decompression,
-    // insertedContent.length would be 5 even though textLengths declares 1.
-    // This is what a decompression-bomb / content-injection payload would
-    // produce, and we want the codec to reject it.
-    const tamperedGraph = new EventGraph();
-    tamperedGraph.addEvent({
-      id: "real:0",
-      parentVersion: new Set(),
-      operation: { type: OPERATION_TYPE.INSERT, index: 0, text: "x" },
-      timestamp: 1,
-    });
-    const tampered = codec.encodeBinary(tamperedGraph);
-
-    // Splice the real payload's LZ4 frame into the tampered payload by
-    // mocking lz4.decompress to return the larger content while textLengths
-    // remains 1. lz4js silently truncates to maxInsertedBytes; the codec's
-    // length-equality check is what catches the tampering.
-    const decompressSpy = vi
-      .spyOn(lz4, "decompress")
-      .mockReturnValue(new TextEncoder().encode("hello"));
-
-    try {
-      expect(() => codec.decodeBinary(tampered)).toThrow(
-        /Decompressed inserted-content size mismatch/,
-      );
-    } finally {
-      decompressSpy.mockRestore();
-    }
-    // Sanity-check the real payload still round-trips with the spy
-    // restored — confirms the tampering was the only thing the test
-    // depended on.
-    expect(codec.decodeBinary(realPayload).getAllEvents()).toHaveLength(1);
-  });
-
-  it("rejects binary payloads whose magic prefix is too short", () => {
-    const codec = new ColumnarEventGraphCodec();
-    // Length-prefix says 3 bytes of magic, but EGW3 is 4 bytes. Even though
-    // the bytes that ARE present match, the length must equal the magic.
-    expect(() =>
-      codec.decodeBinary(new Uint8Array([3, 0x45, 0x47, 0x57])),
-    ).toThrow("Invalid eg-walker columnar graph header");
-  });
-
-  it("rejects binary payloads from older incompatible versions (EGW1, EGW2)", () => {
-    const codec = new ColumnarEventGraphCodec();
-    // 4-byte EGW1 prefix; current decoder expects EGW3.
-    const egw1Header = new Uint8Array([4, 0x45, 0x47, 0x57, 0x31]);
-    expect(() => codec.decodeBinary(egw1Header)).toThrow(
-      "Invalid eg-walker columnar graph header",
-    );
-    // 4-byte EGW2 prefix; also rejected after the EGW3 layout change.
-    const egw2Header = new Uint8Array([4, 0x45, 0x47, 0x57, 0x32]);
-    expect(() => codec.decodeBinary(egw2Header)).toThrow(
-      "Invalid eg-walker columnar graph header",
-    );
-  });
-
-  it("survives deep histories without recursion-stack overflow", () => {
-    const graph = new EventGraph();
-    const total = 25_000;
-    for (let i = 0; i < total; i++) {
-      graph.addEvent({
-        id: `deep:${i}`,
-        parentVersion: i === 0 ? new Set() : new Set([`deep:${i - 1}`]),
-        operation: { type: OPERATION_TYPE.INSERT, index: i, text: "a" },
-        timestamp: i,
-      });
-    }
-
-    const ordered = graph.getTopologicalOrder();
-    expect(ordered).toHaveLength(total);
-    expect(graph.expandVersion(graph.getFrontier()).size).toBe(total);
-
-    const reSerialized = EventGraph.deserialize(graph.serialize());
-    expect(reSerialized.getAllEvents()).toHaveLength(total);
   });
 });
